Update proto enums for supported cointypes
