+++ .vscode/settings.json
@@ -46,5 +46,6 @@
 		"**/.eslintrc.json": true,
 		"**/go.work": true,
 		"**/go.work.sum": true,
+		"**/**.pb.go": true,
 	}
 }
