+++ pkg/did/coins/atom.go
@@ -1,1 +1,6 @@
 package coins
+
+const (
+    ATOM_HRP = "cosmos"
+)
+
