+++ pkg/did/coins/filecoin.go
@@ -1,0 +1,1 @@
+package coins
